refactor(ux-tabs): Refactor how ux-tabs selection transition is animated
