+++ src/tabs/ux-tab-theme.ts
@@ -5,4 +5,5 @@ export class UxTabTheme {
   public foreground: string;
   public background: string;
   public iconSize: string;
+  public indicator: string;
 }
