+++ src/tabs/ux-tabs-theme.ts
@@ -2,5 +2,5 @@ import {styles} from '../styles/decorators';
 
 @styles()
 export class UxTabsTheme {
-  public indicatorColor: string;
+  public indicator: string;
 }
